+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 12.18.0-beta.1 (Unreleased)
 
 ### Features Added
+- Introducing a new field isMultipleClosedAllowed for BlobOutputStream and BlockBlobOutputStream, which allows users to
+  specify if close() will be called on the stream multiple times, and whether to ignore STREAM_CLOSED exception.
 
 ### Breaking Changes